Prepare source code for 64 bit
* Check for sizeof(off_t) > sizeof(size_t)
* Replaced erroneous int type with correct ssize_t type.
* Make textrescompiler support unsigned int/unsigned long.
* Replace "llx" with correct SCNx64
* Replace %d with %zd where appropriate
* textrescompiler generates PRIu64 instead of "llu", ...
* Replace int with unsigned to suppress optimization warnings

+++ C-kern/api/io/ip/ipsocket.h
@@ -188,12 +188,12 @@ int bytestowrite_ipsocket(const ipsocket_t * ipsock, /*out*/size_t * unsend_byte
  * Returns the buffer size in bytes.
  * *readsize* gives the number of bytes which can be received without reading.
  * *writesize* gives the number of bytes which can be written without sending. */
-int queuesize_ipsocket(const ipsocket_t * ipsock, /*out*/uint32_t * readsize, /*out*/uint32_t * writesize);
+int queuesize_ipsocket(const ipsocket_t * ipsock, /*out*/size_t * readsize, /*out*/size_t * writesize);
 
 /* function: setqueuesize_ipsocket
  * Changes the size of the read and write queue.
  * To change only the read or write queue size set the other value to 0. */
-int setqueuesize_ipsocket(ipsocket_t * ipsock, uint32_t queuesize_read/*0 => no change*/, uint32_t queuesize_write/*0 => no change*/);
+int setqueuesize_ipsocket(ipsocket_t * ipsock, size_t queuesize_read/*0 => no change*/, size_t queuesize_write/*0 => no change*/);
 
 // group: connected io
